# improve performance for string range indexing