```jl
using BenchmarkTools
using Random
using BenchmarkTools
s = randstring(200)
f_ib(s) = @inbounds s[1:100]
f(s) = s[1:100]
```

Master

```jl
julia> @btime f($s);
  103.603 ns (1 allocation: 128 bytes)

julia> @btime f_ib($s);
  103.692 ns (1 allocation: 128 bytes)
```

PR

```jl
julia> @btime f($s);
  37.007 ns (1 allocation: 128 bytes)

julia> @btime f_ib($s);
  25.304 ns (1 allocation: 128 bytes)
```

Not sure we have any Nanosoldier for this but might as well

Also added some `@inline` where functions expected themselves to get inlined (since they used `@boundscheck`).

@nanosoldier `runbenchmarks(ALL, vs = ":master")`