@KristofferC 2018-10-25
Made some changes, fire off another one

@nanosoldier `runbenchmarks(ALL, vs = ":master")`