@nanosoldier 2018-10-25
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29796) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/636fc0a965dc1fe5c4fc607583d3c0d54235c579/474809e_vs_44f2563/report.md). cc @ararslan