@nanosoldier 2018-10-25
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29796#issuecomment-432896370) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/8f15771a149906ba0adbe0040d36ee0c997bf5fd/cfc3b6b_vs_86b7e2c/report.md). cc @ararslan